+++ doc.go
@@ -1,6 +1,4 @@
 /*
 Package openfeature provides global access to the OpenFeature API.
-
-Deprecated: use github.com/open-feature/go-sdk, instead.
 */
 package openfeature
