+++ e2e/common_test.go
@@ -1,8 +1,8 @@
 package e2e_test
 
 import (
-	"github.com/open-feature/go-sdk/pkg/openfeature"
-	"github.com/open-feature/go-sdk/pkg/openfeature/memprovider"
+	"github.com/open-feature/go-sdk"
+	"github.com/open-feature/go-sdk/memprovider"
 )
 
 // ctxFunction is a context based evaluation callback
