+++ noop_provider_test.go
@@ -3,7 +3,7 @@ package openfeature_test
 import (
 	"testing"
 
-	"github.com/open-feature/go-sdk/pkg/openfeature"
+	"github.com/open-feature/go-sdk"
 )
 
 func TestNoopProvider_Metadata(t *testing.T) {
